redbpf-probes: refactor network parsing code
Split out the parsing code from XdpContext to a new net::NetworkBuffer
trait, then make XdpContext implement the trait.
